requirements.txt
asgiref==3.2.7
click==7.1.1
Django==3.0.5
joblib==0.14.1
nltk==3.5
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pytesseract==0.3.3
pytz==2019.3
regex==2020.4.4
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
sqlparse==0.3.1
tqdm==4.45.0
